Remove obsolete fragment identifier from Arduino IDE 2.x download link (#111)
During the pre-release development phase of the project, the download links for Arduino IDE 2.x were on a sub-section of
the "Software" page. For this reason, the linked URL included the fragment identifier
`#future-version-of-the-arduino-ide` so that the page would load scrolled down to the anchor at that section near the
bottom of the page.

With the 2.0.0 release, the Arduino IDE 2.x project has graduated to a production development phase. For this reason,
the download links have been moved to the top of the "Software" page and the `future-version-of-the-arduino-ide` anchor
removed from the page.

The previous link with that fragment is still perfectly functional, but the fragment to a non-existent anchor serves no
purpose and also somewhat misrepresents the project status to users who notice the URL that was loaded. For this reason,
it is hereby removed from the link.