Add article re: "bad CPU type in executable" error (#172)
We saw a spike in reports of this error following the recent release of the native Apple Silicon build of Arduino IDE.

Investigation showed that the affected users did not have the Rosetta 2 translation software required to run executables
built for x86 (AKA "Intel") host architecture on Apple Silicon machines.

Previously, this condition was not likely to occurr because macOS shows a dialog when the user starts an x86 application
if Rosetta 2 is not already installed. But it turns out macOS doesn't do that if the x86 executable is run by a
subprocess. So users of the Apple Silicon build of Arduino IDE do not get the dialog when starting Arduino IDE, since
the application itself does not require Rosetta 2, but when Arduino IDE executes one of the helper tools (e.g., avr-gcc)
that is not built for Apple Silicon, the process fails with the "bad CPU in executable" error.

The solution is for the user to manually install Rosetta 2. Instructions for that are provided here.

The article is intended to be a comprehensive guide to solving the "bad CPU type in executable" error. This error may
occur under other conditions than the one described above, so instructions for the resolution of the error as it occurs
under those other conditions are also included.