Document UNO R4 WiFi-specific fix for "No device found on ..." upload error
When an UNO R4 WiFi board is running a sketch that uses the HID capabilities, the port address changes during the
upload. Since this change does not occur under other conditions, the platform is not configured to handle such a change.
This causes uploads via the standard procedure to fail under these conditions:

No device found on <port>

This error can occur under other conditions and with other boards, so there is already an entry for it in the Help
Center. However, the documented fixes are not applicable to the error as it occurs under these special conditions.

There is a reliable fix for the error specific to the UNO R4 WiFi, which was not previously documented in the article.