Adjust UNO R4 WiFi bridge FW restore commands to accommodate any version (#349)
The UNO R4 WiFi bridge firmware packaging system includes the version number in the binary filename. Previously, the
exact filename was referenced in the instructions for restoring the firmware using espflash. This approach meant that
the instructions would need to be updated on every new release of the firmware or else become obsolete. The fact that
the essential updates to the instructions have not been made for the last eight months shows that this approach is not
sustainable.

The commands are hereby adjusted to use a wildcard glob pattern to match the binary filename, avoiding
version-specificity.