Correct information re: platforms path under sketchbook
The Help Center article provides instructions for manual installation of Arduino boards platforms under the sketchbook
folder. Since the sketchbook folder path may differ depending on operating system and configuration of Arduino IDE, the
placeholder `<sketchbook>` is used to refer to the path of the sketchbook folder in the article.

Previously some of the instructions in the article incorrectly included a folder named "Arduino" under the sketchbook
path.